Complete getElementsByClassName.js
Because of the nature of the errors the tests were throwing, I didn't want to push anything until it was obvious everything was working. Discovered while coding that if an element doesn't have a classList, asking element.classList.contains() will not work because you can't run .contains on something that doesn't exist!